Download images from the blog
Once we deal with all of the ways the name can be defined (including how
some posts use utf8 characters), we can launch this script to download all
of the images in the correct tag.

At this point, we can scrape all images that are in a particular category
and use it for model retraining.