Properly create the record file
By adapting create_pascal_tf_record.py to work for classes other than
aeroplanes we can create the actually required data. This assumes the
following organization:
 images
  > class (pose)
     > training
        > images
	records
     > validation
        > images
	records

Also interesting to note: the records that get saved are absolute paths,
so if we extend this in the future we'll want to convert that to relative
paths. In addition all pictures must be jpg files, no png allowed. And in
addition to that, there's a non-zero chance that the labelImg will create
files that have a width of 0, so not sure what that's about.

Once the data is all set here's the command I ran:

python create_pascal_tf_record_custom_class.py --data_dir=images
--custom_class=pose --output_path=output/training_pose.record
--training_type=training --label_map_path=pascal_label_map.pbtxt