Remove cruft
Removed debugging statements and some code that assumed the VOC file
structure. For instance, the xml files directly point to the images with
the folder attribute. Assuming a normal user will keep the xml data and
image data tightly coupled (as I did for my setup) this means we do not need
the image subdirectory parameter.